# Add RPC support to sql/driver.