RPC is significantly faster than HTTP:

```
name         old time/op     new time/op    delta
Select1-8    210us +- 1%     127us +- 1%    -39.54%  (p=0.008 n=5+5)
Select1S-8   232us +- 1%     135us +- 2%    -41.90%  (p=0.008 n=5+5)
```

The "S" suffix is for secure. This is on go1.5. I also tested go-tip and
the improvement is comparable, though the overall perf numbers are about
10% lower on tip across the board (i.e. the go-tip SSL improvements are
being drowned by something else).

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/2994)
<!-- Reviewable:end -->