@petermattis 2015-11-03
I need to add some tests that use the RPC support. Or switch the default.