@tamird 2015-11-03
code LGTM but I don't think we want to go down this road in practice; perhaps we should push the gRPC branch over the finish line instead?