@petermattis 2015-11-03
Do you still have the gRPC branch around? I'd be happy to try benchmarking that as well. And I'm also going to benchmark the pgwire branch.
<!-- Reviewable comment -K2DHSn0-rlS0lW9q-Rw:1079462075 -->

---

Review status: 0 of 10 files reviewed at latest revision, 8 unresolved discussions.

---

<sup>**[sql/bench_test.go, line 70 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jys-r2-70)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/bench_test.go#L70)):</sup>
See my other comment.
<!-- Reviewable comment -K2DHPd2MTkU6IH9SQuh:130912033 -->

---

<sup>**[sql/driver/driver_test.go, line 393 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2DDoBHI8Osm7LLG0Ap-r2-393)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/driver/driver_test.go#L393)):</sup>
DRY is overrated. These constants are very unlikely to change. Even if they were exported from `base/context.go` we'd have to take additional care that a new constant is not added. I struggle to see a net benefit here.
<!-- Reviewable comment -K2DHDR3nCtTWvYnbNSg:887322733 -->

---

<sup>**[sql/driver/driver_test.go, line 404 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2DDoBHI8Osm7LLG0Ap-r2-404)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/driver/driver_test.go#L404)):</sup>
DRY is overrated.
<!-- Reviewable comment -K2DGBSQWxPL6LjZpASB:1882611578 -->

---

<sup>**[sql/driver/rpc_sender.go, line 42 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jyz-r2-42)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/driver/rpc_sender.go#L42)):</sup>
Such little benefit for an additional dependency between packages. And we don't attempt this sharing elsewhere (e.g. https://github.com/cockroachdb/cockroach/blob/master/client/rpc_sender.go#L45), nor do I think we should.
<!-- Reviewable comment -K2DFxw7P9Sc_rJPyWiE:-1085662277 -->

---

<sup>**[sql/server.go, line 34 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jz--r2-34)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/server.go#L34)):</sup>
That's the way goimports formats this stuff. We certainly aren't strict about putting it before cockroach imports elsewhere.
<!-- Reviewable comment -K2DFKG0sjddFJve2-uL:-1172732970 -->

---

<sup>**[testutils/build.go, line 34 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jyy-r2-34)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/testutils/build.go#L34)):</sup>
`cockroach/c-lz4` is being linked into the client and that package returns a no buildable go files error (a.k.a. `build.NoGoError`).
<!-- Reviewable comment -K2DF7W-Ja7GAgSuchVo:-90954304 -->

---


Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994)
<!-- Sent from Reviewable.io -->