@tamird 2015-11-03
I sure do https://github.com/tamird/cockroach/tree/sql-grpc
<!-- Reviewable comment -K2DIkvNCIiGGMt-nb6D:-79564316 -->

---

Review status: 0 of 10 files reviewed at latest revision, 9 unresolved discussions.

---

<sup>**[sql/driver/driver_test.go, line 393 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2DDoBHI8Osm7LLG0Ap-r2-393)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/driver/driver_test.go#L393)):</sup>
I don't think it's overrated. Just yesterday I submitted a PR that fixed benchmarks that had rotted in exactly this way.
<!-- Reviewable comment -K2DIuHtthwZEP2mYOwt:-1929222767 -->

---

<sup>**[sql/driver/rpc_sender.go, line 42 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jyz-r2-42)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/driver/rpc_sender.go#L42)):</sup>
Isn't this dependency already in place?
<!-- Reviewable comment -K2DJ2MbSuTdAAyn5z5y:-1242011164 -->

---

<sup>**[sql/server.go, line 34 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jz--r2-34)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/server.go#L34)):</sup>
goimports is agnostic; it does this by default but if you separate them by a space it will honour your choice. I'm pretty sure we try to organize things elsewhere.
<!-- Reviewable comment -K2DJ6KPCrOuiULm1PS3:1471178632 -->

---

<sup>**[testutils/build.go, line 34 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jyy-r2-34)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/testutils/build.go#L34)):</sup>
That's worth a comment, I think.
<!-- Reviewable comment -K2DJFpKDf0HQqJQcX5y:478835213 -->

---


Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994)
<!-- Sent from Reviewable.io -->