@petermattis 2015-11-03
Review status: 0 of 10 files reviewed at latest revision, 9 unresolved discussions.

---

<sup>**[sql/driver/driver_test.go, line 393 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2DDoBHI8Osm7LLG0Ap-r2-393)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/driver/driver_test.go#L393)):</sup>
Which PR are you referring to? The `client/client_test.go` change was something different. Wasn't it?
<!-- Reviewable comment -K2DYVBSycywkNPDS8jZ:-2067673674 -->

---

<sup>**[sql/driver/rpc_sender.go, line 42 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jyz-r2-42)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/driver/rpc_sender.go#L42)):</sup>
Yeah, the dep is already there. Feels weird for the driver to be defining the name of the endpoint for the server to use and yet it would be silly to introduce an additional package both `sql` and `sql/driver` could point at just to DRY this up.  Changed.
<!-- Reviewable comment -K2DZ3cSzbEGiQD7OKTN:2037941865 -->

---

<sup>**[sql/driver/rpc_sender.go, line 60 \[r3\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jyz-r3-60)** ([raw file](https://github.com/cockroachdb/cockroach/blob/fa08a6fed5a18f5b83a31ea462b4e5db4c470fdd/sql/driver/rpc_sender.go#L60)):</sup>
We don't know about dev mode here. But there is likely cleanup work to be done as fallout from this PR. I think that's ok as we're getting started on performance work.
<!-- Reviewable comment -K2DYitEpHHllj1p5gGa:-103162033 -->

---

<sup>**[sql/server.go, line 34 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jz--r2-34)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/sql/server.go#L34)):</sup>
`git grep` says we're very inconsistent.
<!-- Reviewable comment -K2DYyEYbp5l6ue4EBkT:2069718522 -->

---

<sup>**[testutils/build.go, line 34 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994#-K2D4iii5nU2jYRN_jyy-r2-34)** ([raw file](https://github.com/cockroachdb/cockroach/blob/31e09b8be1eef039809b2bf5cadaa43281a9ae93/testutils/build.go#L34)):</sup>
Sure.
<!-- Reviewable comment -K2DZ-cGH6atJVQ00FOw:-485110604 -->

---


Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/2994)
<!-- Sent from Reviewable.io -->